r11082:
[release] Fix a bunch of CLang compiler warnings

r11095:
LtsView: Fixed some spurious compiler warnings.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/release@11105 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/ltsview/markstateruledialog.cpp
@@ -26,7 +26,7 @@ MarkStateRuleDialog::MarkStateRuleDialog(QWidget *parent, LTS* lts, QColor color
     m_ui.parameterList->addItem(QString::fromStdString(m_lts->getParameterName(i)));
   }
 
-  assert(parameter < m_lts->getNumParameters());
+  assert(parameter < static_cast<int>(m_lts->getNumParameters()));
   m_ui.parameterList->item(parameter)->setSelected(true);
 
   if (negated)
